requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
Django==2.2.24
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==3.3
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
